Conflicts:

	build.xml
	midcom_core/component/baseclass.php
	midcom_core/component/interface.php
	midcom_core/component/loader.php
	midcom_core/configuration/defaults.yml
	midcom_core/controllers/comet.php
	midcom_core/controllers/page.php
	midcom_core/documentation/usage.txt
	midcom_core/exceptionhandler.php
	midcom_core/framework.php
	midcom_core/helpers/context.php
	midcom_core/helpers/head.php
	midcom_core/helpers/navigation.php
	midcom_core/helpers/toolbar/tal.php
	midcom_core/midcom.php
	midcom_core/services/authentication/basic.php
	midcom_core/services/authorization.php
	midcom_core/services/authorization/simple.php
	midcom_core/services/configuration.php
	midcom_core/services/configuration/yaml.php
	midcom_core/services/dispatcher/manual.php
	midcom_core/services/dispatcher/midgard.php
	midcom_core/services/sessioning.php
	midcom_core/services/sessioning/midgard.php
	midcom_core/services/templating/midgard.php
	midcom_core/services/toolbars.php
	midcom_core/services/toolbars/float.php
	midcom_core/services/uimessages.php
	midcom_core/services/uimessages/midgard.php
	midcom_core/static/midcom.js
	midcom_core/static/services/toolbars/float.css
	midcom_core/static/services/toolbars/float.js
	midcom_core/templates/ROOT.php
	midcom_core/templates/midcom-comet-messages.php
	midcom_core/templates/midcom-delete-page.php
	midcom_core/templates/midcom-edit-page.php
	midcom_core/templates/midcom_helpers_toolbar_tal.php
	midcom_helper_datamanager/datamanager.php
	midcom_helper_datamanager/exceptions.php
	midcom_helper_datamanager/manifest.yml
	midcom_helper_datamanager/schema.php
	midcom_helper_datamanager/storage.php
	midcom_helper_datamanager/storage/midgard.php
	midcom_helper_datamanager/tests/typesTest.php
	midcom_helper_datamanager/type.php
	midcom_helper_datamanager/typeproxy.php
	midcom_helper_datamanager/widget.php
	midcom_helper_datamanager/widget/text.php
	midcom_helper_datamanager/widgetproxy.php
	net_nemein_comettest/controllers/index.php
	net_nemein_news/configuration/defaults.yml
	net_nemein_news/configuration/schema.yml
	net_nemein_news/controllers/article.php
	net_nemein_news/templates/nnn-edit-article.php
	net_nemein_news/templates/nnn-show-article.php
	net_nemein_news/tests/articleTest.php
	net_nemein_news/tests/routesTest.php
	org_openpsa_qbpager/pager.php
	scaffold/tests/all.php
	tests/all.php
	tests/helpers.php
	tests/selenium/all.php
	tests/selenium/localhostTest.php
	tests/seleniumcase.php
	tests/testcase.php


git-svn-id: https://svn.midgard-project.org/midgard/trunk/midcom@19573 62f49d1c-1c16-0410-9156-d311ce72abac

+++ midcom_core/services/templating/midgard.php
@@ -20,6 +20,8 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
 
     private $elements_shown = array();
 
+    private $gettext_translator = array();
+
     public function __construct()
     {
         $this->stacks[0] = array();
@@ -27,7 +29,13 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
     
     private function get_cache_identifier()
     {
-        return "{$_MIDGARD['host']}-{$_MIDGARD['page']}-{$_MIDGARD['style']}-" . $_MIDCOM->context->get_current_context() . "-{$_MIDCOM->context->route_id}";
+        if (isset($_MIDCOM->context->route_id))
+        {
+            return "{$_MIDGARD['host']}-{$_MIDGARD['page']}-{$_MIDGARD['style']}-" . $_MIDCOM->context->get_current_context() . 
+                   "-{$_MIDCOM->context->route_id}-{$_MIDCOM->context->template_entry_point}-{$_MIDCOM->context->content_entry_point}";
+        }
+        return "{$_MIDGARD['host']}-{$_MIDGARD['page']}-{$_MIDGARD['style']}-" . $_MIDCOM->context->get_current_context() . 
+               "-{$_MIDCOM->context->template_entry_point}-{$_MIDCOM->context->content_entry_point}";
     }
     
     private function get_cache_directory()
@@ -52,7 +60,7 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
         
         if (!is_writable($this->cache_directory))
         {
-            throw new Exception("Cache directory {$cache_directory} is not writable");
+            throw new Exception("Cache directory {$this->cache_directory} is not writable");
         }
     }
     
@@ -362,11 +370,16 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
     {
         $_MIDCOM->context->create();
         $data = $this->dynamic_call($component_name, $route_id, $arguments, false);
-        
+
         $this->template('content_entry_point');
         $this->display();
-        
+
+        /* 
+         * Gettext is not context safe. Here we return the "original" textdomain
+         * because in dynamic call the new component may change it
+         */
         $_MIDCOM->context->delete();
+        $_MIDCOM->i18n->set_translation_domain($_MIDCOM->context->component_name);
     }
 
     /**
@@ -374,9 +387,12 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
      */    
     public function template($element_identifier = 'template_entry_point')
     {
+        // Let injectors do their work
+        $_MIDCOM->componentloader->inject_template();
+
         $this->prepare_cache();
         $cache_file = $this->cache_directory . '/' . $this->get_cache_identifier() . '.php';
-        
+
         if (file_exists($cache_file))
         {
             return;
@@ -398,18 +414,23 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
         ob_start();
         include($cache_file);
         $content = ob_get_clean();
-        // FIXME: Remove this once we can actually invalidate cache
-        unlink($cache_file);
-
+        
+        if (!$_MIDCOM->configuration->templating_cache)
+        {
+            // FIXME: Remove this once we can actually invalidate cache
+            unlink($cache_file);
+        }
         switch ($data['template_engine'])
         {
             case 'tal':
                 if (!class_exists('PHPTAL'))
                 {
                     require('PHPTAL.php');
-                    include_once('TAL/modifiers.php');
                 }
-                
+
+                // FIXME: Rethink whole tal modifiers concept 
+                include_once('TAL/modifiers.php');
+
                 if ($_MIDCOM->timer)
                 {
                     $_MIDCOM->timer->setMarker('post-require');
@@ -427,14 +448,15 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
                 if ($_MIDCOM->timer)
                 {
                     $_MIDCOM->timer->setMarker('post-set-show_toolbar');
-                }
-                
+                }              
                 $tal->uimessages = false;
-                $uimessages = new midcom_core_services_uimessages();
-                if (   $uimessages->has_messages()
-                    && $uimessages->can_view())
+                if ($_MIDCOM->configuration->enable_uimessages)
                 {
-                    $tal->uimessages = $uimessages->render();
+                    if (   $_MIDCOM->uimessages->has_messages()
+                        && $_MIDCOM->uimessages->can_view())
+                    {
+                        $tal->uimessages = $_MIDCOM->uimessages->render();
+                    }
                 }
 
                 if ($_MIDCOM->timer)
@@ -468,7 +490,7 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
                         $_MIDCOM->timer->setMarker("post-set-{$key}");
                     }
                 }
-                
+
                 $tal->setSource($content);
                 
                 if ($_MIDCOM->timer)
@@ -476,7 +498,11 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
                     $_MIDCOM->timer->setMarker('post-source');
                 }
                 
+                $translator =& $_MIDCOM->i18n->set_translation_domain($_MIDCOM->context->component);
+                $tal->setTranslator($translator);  
+                          
                 $content = $tal->execute();
+                unset($tal);
                 
                 if ($_MIDCOM->timer)
                 {
@@ -507,10 +533,13 @@ class midcom_core_services_templating_midgard implements midcom_core_services_te
             }
             echo "</ul>\n";
         }
-        
-        ///TODO: Connect this to some signal that tells the MidCOM execution has ended.
-        $uimessages = new midcom_core_services_uimessages();
-        $uimessages->store();
+
+        if ($_MIDCOM->configuration->enable_uimessages)
+        {
+            ///TODO: Connect this to some signal that tells the MidCOM execution has ended.
+            $_MIDCOM->uimessages->store();
+        }
     }
+     
 }
 ?>
